requirements.txt
Django>=4.0
djangorestframework
celery
redis
psycopg2-binary
python-dotenv
mysqlclient
python-decouple
django-redis
requests
drf-yasg
pytest
